pyproject.toml
[project]
name = "osv"
version = "0.0.22"
description = "Open Source Vulnerabilities library"
authors = [
    {name = "OSV Authors", email = "osv-discuss@googlegroups.com"}
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "google-cloud-ndb>=2.3", 
    "google-cloud-logging>=3.10", 
    "google-cloud-storage>=2.17", 
    "semver>=3.0",
    "packageurl-python==0.16.0",
    "pyyaml>=6.0",  
    "pygit2>=1.14.0", 
    "attrs>=23.2", 
    "jsonschema>=4.0", 
    "grpcio>=1.0", 
    "packaging-legacy>=23.0.post0",
    "requests>=2.32", 
]


[tool.poetry.group.dev.dependencies]
yapf = "*"
pylint = "*"
grpcio-tools = "*"
mypy-protobuf = "^3.6.0"
vcrpy = "*"

[tool.pyright]
executionEnvironments = [
    { root = "osv/models.py", reportAssignmentType = "none" },
]

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
